refactor Shoppinglist page and consider refacalculated recipes in it
