+++ lib/helper.js
@@ -72,9 +72,15 @@ export async function uploadRecipeStepImages(recipeSteps) {
   );
 }
 
-export function calculateShoppingList(recipes) {
-  const allIngredients = recipes.flatMap((recipe) => recipe.ingredients);
-
+export function calculateShoppingList(recipes, recalculatedRecipes) {
+  const allIngredients = recipes.flatMap((recipe) => {
+    const scalingFactor = recalculatedRecipes[recipe._id]?.scalingFactor ?? 1;
+    return recipe.ingredients.map(({ ingredient, amount, unit }) => ({
+      ingredient,
+      amount: Math.round(amount * scalingFactor * 100) / 100,
+      unit,
+    }));
+  });
   const merged = new Map();
 
   allIngredients.forEach(({ ingredient, amount, unit }) => {
